Simplify with translate method within spec data class
