+++ src/main/java/com/rackspace/salus/monitor_management/web/model/MonitorTranslationOperatorCreate.java
@@ -16,7 +16,7 @@
 
 package com.rackspace.salus.monitor_management.web.model;
 
-import com.rackspace.salus.monitor_management.web.model.translators.MonitorTranslatorSpec;
+import com.rackspace.salus.monitor_management.web.model.translators.MonitorTranslator;
 import com.rackspace.salus.telemetry.model.AgentType;
 import com.rackspace.salus.telemetry.model.ConfigSelectorScope;
 import com.rackspace.salus.telemetry.model.MonitorType;
@@ -48,5 +48,5 @@ public class MonitorTranslationOperatorCreate {
   ConfigSelectorScope selectorScope;
 
   @NotNull @Valid
-  MonitorTranslatorSpec translatorSpec;
+  MonitorTranslator translatorSpec;
 }
