+++ src/main/java/com/rackspace/salus/monitor_management/web/model/MonitorTranslationOperatorDTO.java
@@ -17,7 +17,7 @@
 package com.rackspace.salus.monitor_management.web.model;
 
 import com.rackspace.salus.monitor_management.entities.MonitorTranslationOperator;
-import com.rackspace.salus.monitor_management.web.model.translators.MonitorTranslatorSpec;
+import com.rackspace.salus.monitor_management.web.model.translators.MonitorTranslator;
 import com.rackspace.salus.telemetry.model.AgentType;
 import com.rackspace.salus.telemetry.model.ConfigSelectorScope;
 import com.rackspace.salus.telemetry.model.MonitorType;
@@ -37,7 +37,7 @@ public class MonitorTranslationOperatorDTO {
 
   ConfigSelectorScope selectorScope;
 
-  MonitorTranslatorSpec translatorSpec;
+  MonitorTranslator translatorSpec;
 
   public MonitorTranslationOperatorDTO(MonitorTranslationOperator entity) {
     this.id = entity.getId();
